Fix check_eq when comparing values (no json data)
The goal was to compare values out of a json array:
example data:
[
    {
        "key": "value",
        "foo": "bar",
        "status": "ASTATUS"
    }
]

example test case:
    ...
    "testCases": {
        "get_status": {
            "path": "/programs/",
            "method": "GET",
            "description": "GET /programs",
            "expect": {
                "body": {
                    "path_eq": {
                        "[0].status": "MYSTATUS"
                    }
                }
            }
        }
    },

This causes 'check_eq' to compare "ASTATUS" (from the data) with
"MYSTATUS" from the test description. This is what I want.
However the given arguements are not json data, and jq fails to retrieve
its' type and eventually fails to make a valid comparison.

First off all the type determination is expected to fail, which is
hidden with the redirection of stderr.

The actual type-comparisons are extended with all recognized types, and
everything else is just compared as a string.